+++ CHANGES.md
@@ -15,3 +15,6 @@
 * added option to transform email mime parts
 * fixed postmark to_string
 * added parse_restore option to generate N unique messages from an archive
+* update for ocaml-tls 0.4.0
+* store message parts keyed by the message hash and part number 0-post;1-headers;2-content;3+-attachments. All messages are stored under imaplet:user:storage key
+* the message is staged/commited under View keyed off imaplet:user